backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
mysqlclient==2.2.4
pydantic==2.7.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
faker==25.2.0
sqlalchemy_utils==0.41.2
pymysql==1.1.1
cryptography==42.0.7
httpx==0.27.0
requests==2.32.5
